forest_fire_backend/pyproject.toml
[project]
name = "forest-fire-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.0.0,<4.1.0",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "opencv-python>=4.13.0.92",
    "passlib[bcrypt]>=1.7.4",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "ultralytics>=8.4.21",
    "uvicorn[standard]>=0.41.0",
]
